Move envelope tests to GeometryFunctionsTest
